pyproject.toml
[project]
name = "tidb-lab-branches"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "graphviz>=0.21",
    "ipywidgets>=8.1.8",
    "nbformat>=5.10.4",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "pytidb>=0.0.13",
    "requests>=2.32.5",
    "strands-agents>=1.17.0",
    "streamlit>=1.51.0",
]
